optimized PWM generation
switched PWM calculations to integers

+++ gimbal-motor/src/lib/inc/bldc.hpp
@@ -17,8 +17,8 @@
 namespace bldc {
     void init();
     
-    // Angle from 0 to 3600 (1/10th of a degree), power from 0 to 100
-    void applyTorque(uint16_t angle, uint16_t power = 100);
+    // Angle from 0 to 4096 (0 = 0deg, 4096 = 360deg), power from 0 to 255
+    void applyTorque(uint16_t angle, uint8_t power = 100);
     void tone(uint16_t frequency);
     void silent();
 }
